ai-service/requirements.txt
fastapi
uvicorn
requests
langchain
langgraph
langchain-openai
dotenv
pydantic
langchain_groq
langchain_core
sqlalchemy
psycopg2-binary
passlib[bcrypt]
python-jose
argon2-cffi